add redirect to GalleryBlock
