+++ resources/assets/components/utilities/ScholarshipCard/ScholarshipCard.js
@@ -2,19 +2,12 @@ import React from 'react';
 import gql from 'graphql-tag';
 import { propType } from 'graphql-anywhere';
 
-import SecondaryButton from '../Button/SecondaryButton';
 import GalleryBlockSignup from './GalleryBlockSignup';
 import {
   contentfulImageSrcset,
   contentfulImageUrl,
   getHumanFriendlyDate,
 } from '../../../helpers';
-import {
-  EVENT_CATEGORIES,
-  trackAnalyticsEvent,
-  getPageContext,
-} from '../../../helpers/analytics';
-import { isAuthenticated } from '../../../helpers/auth';
 
 // Write a graphql query to get campaign information for a specific id
 export const scholarshipCardFragment = gql`
@@ -58,19 +51,6 @@ const ScholarshipCard = ({ campaign }) => {
     { height: 410, width: 730 },
     { height: 820, width: 1460 },
   ]);
-
-  const handleScholarshipCardShareClick = () => {
-    trackAnalyticsEvent('clicked_scholarship_gallery_block_apply_now', {
-      action: 'button_clicked',
-      category: EVENT_CATEGORIES.siteAction,
-      label: 'scholarship_gallery_card',
-      context: {
-        url: path,
-        ...getPageContext(),
-        campaignId,
-      },
-    });
-  };
 
   return (
     <article className="flex flex-col h-full relative text-left">
@@ -118,16 +98,7 @@ const ScholarshipCard = ({ campaign }) => {
           </div>
         </div>
 
-        {isAuthenticated() ? (
-          <GalleryBlockSignup campaignId={campaignId} path={path} />
-        ) : (
-          <SecondaryButton
-            className="w-full"
-            href={path}
-            text="Apply Now"
-            onClick={handleScholarshipCardShareClick}
-          />
-        )}
+        <GalleryBlockSignup campaignId={campaignId} path={path} />
       </div>
     </article>
   );
